Move cache up so that its shared between prefetch threads
