+++ crates/kellnr/Cargo.toml
@@ -27,6 +27,7 @@ axum-extra.workspace = true
 tower-http.workspace = true
 tokio.workspace = true
 openssl = { version = "*", optional = true } # Not needed directly but for cross-compilation with the vendored-openssl feature
+moka.workspace = true
 
 [features]
 vendored-openssl = ["openssl/vendored"]
